Attribution note and update contact for fork
I'm forking this project in order to use if in my own courses. Most of the work was done by puppetlabs, including some commits that I made when I worked there, and that should all be attributed to them. I no longer work at puppetlabs so any changes after this point are either mine, or code merged from the community.